Добавил текст: Неверный статус кода

+++ api.py
@@ -11,7 +11,7 @@ app.mount("/static", StaticFiles(directory="public"))
 @app.get("/kv_ur")
 def kv_ur_get(a, b, c):
     """эндпоинт для нахождения корней квадратного уравнения"""
-    a = int(a)
-    b = int(b)
-    c = int(c)
+    a = float(a)
+    b = float(b)
+    c = float(c)
     return kv_ur(a, b, c)
